Separate SwarmTest.Broadcast
